requirements.txt
# Secure Chat Application Dependencies
# Signal-inspired E2E encrypted messaging system

# Core cryptographic libraries
cryptography>=41.0.0
bcrypt>=4.0.0

# Web framework and real-time communication
Flask>=2.3.0
Flask-SocketIO>=5.3.0
python-socketio[client]>=5.8.0

# Authentication and security
PyJWT>=2.8.0

# HTTP client for API calls
requests>=2.31.0

# Database support (SQLite is built-in to Python)
# No additional database packages needed

# Development and testing dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: Enhanced performance and additional features
# eventlet>=0.33.0      # For improved WebSocket performance
# gevent>=23.7.0        # Alternative async backend
# python-dotenv>=1.0.0  # For environment configuration 